DFE-1156 - tidied up page component to separate out into page, form, and summary views separately
